Constrained Delaunay triangulation working
There is a test scene that contains a Triangulation object with a DelaunayTriangulationTester script.
The tester can be filled with 2 sets of points, the point cloud to triangulate and the constrained edges to add afterwards. Use the Tun button to draw the triangles in the Editor scene view. All of the test cases available in the scene are working perfectly.
The project includes a csc.rsp file, to enable unsafe code (required for fixed size arrays).
It is pending to cut off the supertriangle.